remove InstallPlugin method from API
- do not allow InstallPlugin method to be called via API
- move InstallPlugin functionality to PluginsManager for use exclusively
